Rename chart package to cnr
